moved the coordinate assignement to its own file This splits the long function also up into smaller chuks for readability
